refactor: consolidate integration-test infrastructure into single file
Remove duplicated test-container setup and helper code by deleting
integration_infrastructure_test.go and moving all shared helpers
into integration_test.go. This eliminates the need for a separate
README_INTEGRATION_TESTS.md, which is also removed.

The change keeps the same test behaviour while reducing maintenance
surface and ensuring a single source of truth for test setup.
